// lib/generated/prisma/models/JamKerja.ts
export type JamKerjaModel =
    runtime.Types.Result.DefaultSelection<Prisma.$JamKerjaPayload>;
// ...
export type $JamKerjaPayload<
    ExtArgs extends
        runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs,
> = {
    name: "JamKerja";
    objects: {};
    scalars: runtime.Types.Extensions.GetPayloadResult<
        {
            id: number;
            kodeShift: string;
            jamMasuk: Date;
            jamPulang: Date;
            createdAt: Date;
            updatedAt: Date;
        },
        ExtArgs["result"]["jamKerja"]
    >;
    composites: {};
};